backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pysqlite3-binary==0.5.4
python-multipart==0.0.6
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0
